Create a grunt task for generating a parser
